requirements.txt
oslo.config